Inline tsconfig target version
