+++ packages/nx-firebase/src/utils/versions.ts
@@ -7,19 +7,3 @@ export const pluginNxVersionMajor = parseInt(packageVersions.nx.split('.')[0])
 // Target node 16 for all firebase projects now
 export const firebaseNodeEngine = packageVersions.nodeEngine
 export const firebaseNodeRuntime = `nodejs${firebaseNodeEngine}`
-
-// https://stackoverflow.com/questions/59787574/typescript-tsconfig-settings-for-node-js-12
-
-export const nodeEsVersion: Record<string, string> = {
-  '12': 'es2019',
-  '14': 'es2020',
-  '16': 'es2020', // es2020 seems more preferred with node 16 than es2021
-  '18': 'es2022',
-  '20': 'es2022',
-}
-
-export const tsConfigTarget = nodeEsVersion[firebaseNodeEngine]
-
-if (!tsConfigTarget) {
-  throw new Error('Undefined tsConfigTarget')
-}
